refactor: remove deprecated admin actions, loading component, and localization files; update import paths for actions
